+++ esbuild/build.mjs
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-import {build} from 'esbuild';
+import {build} from '@diplodoc/lint/esbuild';
 
 import tsconfigJson from '../tsconfig.json' with {type: 'json'};