+++ CHANGELOG.md
@@ -1,5 +1,83 @@
 # Changelog
 
+## [1.6.0](https://github.com/diplodoc-platform/yfmlint/compare/v1.5.0...v1.6.0) (2026-03-23)
+
+
+### Features
+
+* Change yfm009 to error, add yfm018 ([3a1b63f](https://github.com/diplodoc-platform/yfmlint/commit/3a1b63f9f76bb59b8d3b73f9132821a5eb0bd1e0))
+* Expose getLogLevel helper ([0b0c4c3](https://github.com/diplodoc-platform/yfmlint/commit/0b0c4c30b7ff842b4125410519f4811c1101d83a))
+* Expose normalizeConfig util ([31f0746](https://github.com/diplodoc-platform/yfmlint/commit/31f0746086859bf099a61326e69d48c8c17a43d1))
+* **yfmlint:** use token source metadata for include chain in error context ([392c5b2](https://github.com/diplodoc-platform/yfmlint/commit/392c5b272bb14a76f3f7ba8e7a71a0ea2680553a))
+
+
+### Bug Fixes
+
+* add autotitle anchor rule ([fef7c96](https://github.com/diplodoc-platform/yfmlint/commit/fef7c965c4bed0de2ecbf6bd1f0b8d649ffa2175))
+* add rule for max inline svg size ([aea1e3a](https://github.com/diplodoc-platform/yfmlint/commit/aea1e3a97d4ae2a7fc8277d5c9eaa68ba2f4397d))
+* Change level to loglevel ([419f512](https://github.com/diplodoc-platform/yfmlint/commit/419f512f34502daaaed9810e11e9d6ad6a0b5f15))
+* configure ESLint to properly handle config files and build directory ([a015d67](https://github.com/diplodoc-platform/yfmlint/commit/a015d6757b2e8d26d0550224020562f181f8664d))
+* error in lineNumber DOCSTOOLS-5294 ([6168924](https://github.com/diplodoc-platform/yfmlint/commit/61689247a8bdf560357f8bf49c5038c0820edeca))
+* Fix deps ([329716c](https://github.com/diplodoc-platform/yfmlint/commit/329716ccf9061023971954c9947af84bcfe2b7b6))
+* fix package config ([7a2e3ad](https://github.com/diplodoc-platform/yfmlint/commit/7a2e3ad66e176bdeeb981d4d86c3d1597d8add45))
+* Fix readme and ysage ([69fedd8](https://github.com/diplodoc-platform/yfmlint/commit/69fedd8320a8c16112e3a24c9ea3d3adf1dbc77d))
+* fix reason description in yfm003 ([d392818](https://github.com/diplodoc-platform/yfmlint/commit/d392818c17c2e84127e58363f8a7eb85864f544f))
+* fix reason description in yfm003 ([a2e140a](https://github.com/diplodoc-platform/yfmlint/commit/a2e140a3ca8d75bf6435c773b637ce4bb9088fea))
+* Fix types ([b186a11](https://github.com/diplodoc-platform/yfmlint/commit/b186a11fca760ed81acc948834e64aad537f48cb))
+* fixed line number ([d29d30c](https://github.com/diplodoc-platform/yfmlint/commit/d29d30c301b806d7dbfae0c667a844b44fd8b705))
+* fixed line number ([524fc1f](https://github.com/diplodoc-platform/yfmlint/commit/524fc1fef2190a48def947e49912b7489a77d5a8))
+* Lot of architecture fixes ([9320f11](https://github.com/diplodoc-platform/yfmlint/commit/9320f113e2664f514a69eba07825f17ab4fdf1dc))
+* properly configure ESLint to ignore build directory and config files ([2a6aac8](https://github.com/diplodoc-platform/yfmlint/commit/2a6aac83cb182f0cd180a5299f64e6d64b1b5525))
+* remove console.log statement ([878ba7d](https://github.com/diplodoc-platform/yfmlint/commit/878ba7d53295725d132b24992f8ee2b7b01ea62e))
+* update .eslintignore to properly exclude build directory ([7147d9b](https://github.com/diplodoc-platform/yfmlint/commit/7147d9bb64c87da5ef770534940b7441cbce4014))
+* Update api and tests ([8dbeaf8](https://github.com/diplodoc-platform/yfmlint/commit/8dbeaf80a73f01362bee10781398cecec8ffad86))
+* Update config.ts ([c97ff47](https://github.com/diplodoc-platform/yfmlint/commit/c97ff4746a4059221a89c8b97b786fc8187e6089))
+* Update infra to v1.13.2 ([7418bb5](https://github.com/diplodoc-platform/yfmlint/commit/7418bb55cf730519d8eb2e53b49e2f0eaa8fef32))
+* Update readme ([6ffc4ca](https://github.com/diplodoc-platform/yfmlint/commit/6ffc4cac8f364d9361f6c231845573d50005f093))
+* update yfm003 error context ([c72879b](https://github.com/diplodoc-platform/yfmlint/commit/c72879bdf7329d2062f44816d2c85bf52d76a353))
+* use correct token types in YFM009 rule ([124c600](https://github.com/diplodoc-platform/yfmlint/commit/124c600c3147a5121a00ff266c2cc55ffe64cd9c))
+* Use unstable markdownlint with perf fix ([d514de6](https://github.com/diplodoc-platform/yfmlint/commit/d514de687507f066bb644e79797641d610b38058))
+* warning in YFM003 ([165e0e3](https://github.com/diplodoc-platform/yfmlint/commit/165e0e3ca63adc1e4ddf73354457c95161767687))
+* warning in YFM003 ([0ff43d6](https://github.com/diplodoc-platform/yfmlint/commit/0ff43d63b5a5374a79e58615a452f471606065ae))
+* warning in YFM003 ([73a0682](https://github.com/diplodoc-platform/yfmlint/commit/73a06822cf615ac64a52a666b7b3c990e3c88e9d))
+* warning in YFM003 ([5ef4da1](https://github.com/diplodoc-platform/yfmlint/commit/5ef4da1df6771043f913a7ffcea4be4255d756cc))
+* yfm003 error -&gt; warn ([b8b280c](https://github.com/diplodoc-platform/yfmlint/commit/b8b280c7e4e2231dbd94a9cc86e8d81291c3182c))
+* YFM010 error -&gt; warn ([6791b93](https://github.com/diplodoc-platform/yfmlint/commit/6791b93facef5d7118f8436f52eef143a4768bc8))
+
+## [1.5.0](https://github.com/diplodoc-platform/yfmlint/compare/v1.4.0...v1.5.0) (2026-03-23)
+
+
+### Features
+
+* Change yfm009 to error, add yfm018 ([3a1b63f](https://github.com/diplodoc-platform/yfmlint/commit/3a1b63f9f76bb59b8d3b73f9132821a5eb0bd1e0))
+
+## [1.4.0](https://github.com/diplodoc-platform/yfmlint/compare/v1.3.5...v1.4.0) (2026-03-05)
+
+
+### Features
+
+* **yfmlint:** use token source metadata for include chain in error context ([392c5b2](https://github.com/diplodoc-platform/yfmlint/commit/392c5b272bb14a76f3f7ba8e7a71a0ea2680553a))
+
+## [1.3.5](https://github.com/diplodoc-platform/yfmlint/compare/v1.3.4...v1.3.5) (2026-02-23)
+
+
+### Bug Fixes
+
+* error in lineNumber DOCSTOOLS-5294 ([6168924](https://github.com/diplodoc-platform/yfmlint/commit/61689247a8bdf560357f8bf49c5038c0820edeca))
+* Update infra to v1.13.2 ([7418bb5](https://github.com/diplodoc-platform/yfmlint/commit/7418bb55cf730519d8eb2e53b49e2f0eaa8fef32))
+
+## [1.3.4](https://github.com/diplodoc-platform/yfmlint/compare/v1.3.3...v1.3.4) (2026-01-12)
+
+
+### Bug Fixes
+
+* configure ESLint to properly handle config files and build directory ([a015d67](https://github.com/diplodoc-platform/yfmlint/commit/a015d6757b2e8d26d0550224020562f181f8664d))
+* Fix deps ([329716c](https://github.com/diplodoc-platform/yfmlint/commit/329716ccf9061023971954c9947af84bcfe2b7b6))
+* properly configure ESLint to ignore build directory and config files ([2a6aac8](https://github.com/diplodoc-platform/yfmlint/commit/2a6aac83cb182f0cd180a5299f64e6d64b1b5525))
+* remove console.log statement ([878ba7d](https://github.com/diplodoc-platform/yfmlint/commit/878ba7d53295725d132b24992f8ee2b7b01ea62e))
+* update .eslintignore to properly exclude build directory ([7147d9b](https://github.com/diplodoc-platform/yfmlint/commit/7147d9bb64c87da5ef770534940b7441cbce4014))
+* use correct token types in YFM009 rule ([124c600](https://github.com/diplodoc-platform/yfmlint/commit/124c600c3147a5121a00ff266c2cc55ffe64cd9c))
+
 ## [1.3.3](https://github.com/diplodoc-platform/yfmlint/compare/v1.3.2...v1.3.3) (2025-12-04)